requirements.txt
bcrypt==3.2.0
boto3==1.24.60
Flask==2.0.2
Flask-Admin==1.6.0
Flask-Cors==3.0.10
Flask-Mail==0.9.1
Flask-SocketIO==5.3.0
Flask-SQLAlchemy==2.5.1
eventlet==0.30.2
gevent==21.12.0
greenlet==1.1.2
gunicorn==20.1.0
numpy==1.21.3
matplotlib==3.4.3
pandas==1.3.4
plotly==5.3.1
psycopg2==2.9.3
PyJWT==2.3.0
python-decouple==3.6
scipy==1.9.1